move list item composable to common module
